+++ app/lib/frontend/templates/views/pkg/index.dart
@@ -154,8 +154,8 @@ d.Node _searchFormContainer({
               ),
               _tagBasedCheckbox(
                 tagPrefix: 'is',
-                tagValue: 'dart3-ready',
-                label: 'Dart 3 ready',
+                tagValue: 'dart3-compatible',
+                label: 'Dart 3 compatible',
                 searchForm: searchForm,
                 title: 'Show only packages compatible with Dart 3.',
               ),